move component variables to kea and funnel utility methods into separate file
