+++ frontend/src/scenes/funnels/FunnelBarGraph.tsx
@@ -12,11 +12,7 @@ import { useThrottledCallback } from 'use-debounce'
 import './FunnelBarGraph.scss'
 import { useActions, useValues } from 'kea'
 import { FunnelBarLayout } from 'lib/constants'
-import { FunnelStepReference } from 'scenes/insights/InsightTabs/FunnelTab/FunnelStepReferencePicker'
-
-export function calcPercentage(numerator: number, denominator: number): number {
-    return (numerator / denominator) * 100 || 0
-}
+import { calcPercentage, getReferenceStep } from './funnelUtils'
 
 function humanizeOrder(order: number): number {
     return order + 1
@@ -200,21 +196,6 @@ function AverageTimeInspector({ onClick, disabled, averageTime }: AverageTimeIns
             </ValueInspectorButton>
         </div>
     )
-}
-
-export function getReferenceStep(steps: FunnelStep[], stepReference: FunnelStepReference, index?: number): FunnelStep {
-    // Step to serve as denominator of percentage calculations.
-    // step[0] is full-funnel conversion, previous is relative.
-    if (!index || index <= 0) {
-        return steps[0]
-    }
-    switch (stepReference) {
-        case FunnelStepReference.previous:
-            return steps[index - 1]
-        case FunnelStepReference.total:
-        default:
-            return steps[0]
-    }
 }
 
 export function FunnelBarGraph({ steps: stepsParam }: FunnelBarGraphProps): JSX.Element {
