+++ frontend/src/scenes/funnels/FunnelHistogram.tsx
@@ -3,49 +3,30 @@ import { useActions, useValues } from 'kea'
 import { humanFriendlyDuration, humanizeNumber } from 'lib/utils'
 import React from 'react'
 import { LineGraph } from 'scenes/insights/LineGraph'
-import { calcPercentage, getReferenceStep } from './FunnelBarGraph'
+import { calcPercentage, getReferenceStep } from './funnelUtils'
 import { funnelLogic } from './funnelLogic'
 
-interface TimeStepOption {
-    label: string
-    value: number
-    average_conversion_time: number
-    count: number
-}
-
 export function FunnelHistogram(): JSX.Element {
-    const { timeConversionBins, stepsWithCount, stepReference } = useValues(funnelLogic)
+    const { stepsWithCount, stepReference, histogramGraphData, histogramStepsDropdown } = useValues(funnelLogic)
     const { changeHistogramStep } = useActions(funnelLogic)
-    const labels = timeConversionBins.map((bin) => humanFriendlyDuration(`${bin[0]}`))
-    const binData = timeConversionBins.map((bin) => bin[1])
-    const dataset = [{ data: binData, labels: labels, label: 'Time to convert', count: 3 }]
-
-    const stepsDropdown: TimeStepOption[] = []
-    stepsWithCount.forEach((_, idx) => {
-        if (stepsWithCount[idx + 1]) {
-            stepsDropdown.push({
-                label: `Steps ${idx + 1} and ${idx + 2}`,
-                value: idx + 1,
-                count: stepsWithCount[idx + 1].count,
-                average_conversion_time: stepsWithCount[idx + 1].average_conversion_time,
-            })
-        }
-    })
+    const dataset = [
+        { data: histogramGraphData.personsAmount, labels: histogramGraphData.time, label: 'Time to convert' },
+    ]
 
     return (
         <>
             <div>
                 Steps
-                {stepsDropdown.length > 0 && (
+                {histogramStepsDropdown.length > 0 && (
                     <Select
-                        defaultValue={stepsDropdown[0]?.value}
+                        defaultValue={histogramStepsDropdown[0]?.value}
                         onChange={changeHistogramStep}
                         dropdownMatchSelectWidth={false}
                         data-attr="funnel-bar-layout-selector"
                         optionLabelProp="label"
                         style={{ marginLeft: 8, marginBottom: 16 }}
                     >
-                        {stepsDropdown.map((option, i) => {
+                        {histogramStepsDropdown.map((option, i) => {
                             const basisStep = getReferenceStep(stepsWithCount, stepReference, i)
                             return (
                                 <Select.Option
@@ -76,7 +57,7 @@ export function FunnelHistogram(): JSX.Element {
                 type="bar"
                 color={'white'}
                 datasets={dataset}
-                labels={labels}
+                labels={histogramGraphData.time}
                 dashboardItemId={null}
             />
         </>
